Split events.spec.js into separate files
Breaking it up per component makes it easier to manage.
